Use the same approach as tile selection to display signals
I'm not really pleased with the fact I had to disable the
`display_tile_overlay` system when using the signal overlay but this
can be cleanup later.

Color scheme has been updated to look better with the new terrain and
shadows, and the overlay material has been made transparent a bit.